Выносит логику в модули
