--- js/service/node-from-template.js
@@ -1,8 +1,0 @@
-const getElementFromTemplate = (screenTemplate) => {
-  const newScreen = document.createElement(`section`);
-  newScreen.className = `screen`;
-  newScreen.innerHTML = screenTemplate;
-  return newScreen;
-};
-
-export default getElementFromTemplate;
