+++ js/start.js
@@ -1,0 +1,8 @@
+import welcomeNode from "./templates/welcome";
+import {renderScreen} from "./service";
+
+const startNewGame = () => {
+  renderScreen(welcomeNode);
+};
+
+export default startNewGame;
